requirements.txt
# Core dependencies
torch>=1.10.0  # PyTorch is needed for tensor operations and model handling
transformers  # Provides access to pre-trained transformer models and utilities
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git  # Unsloth library with colab-new extras
xformers<0.0.26  # Transformer extensions, used for Flash Attention and other optimizations
trl  # Transformer Reinforcement Learning, used for training with reinforcement learning techniques
peft  # Parallel Execution for Transformers, optimizes transformer execution
accelerate  # Simplifies running PyTorch models on multi-GPU/multi-TPU setups
bitsandbytes  # For efficient memory usage during model training
datasets  # Used for loading and processing datasets from Hugging Face
psutil  # Allows checking system resources like RAM
python-dotenv  # For loading environment variables from .env file
